Moving `FieldTypesMappingList` type to separate module.
