+++ graylog2-web-interface/src/views/components/datatable/DataTableEntry.tsx
@@ -26,7 +26,7 @@ import { AdditionalContext } from 'views/logic/ActionContext';
 import type { ValuePath } from 'views/logic/valueactions/ValueActionHandler';
 import type Series from 'views/logic/aggregationbuilder/Series';
 import { parseSeries } from 'views/logic/aggregationbuilder/Series';
-import type { FieldTypeMappingsList } from 'views/stores/FieldTypesStore';
+import type { FieldTypeMappingsList } from 'views/logic/fieldtypes/types';
 import fieldTypeFor from 'views/logic/fieldtypes/FieldTypeFor';
 
 import type { CurrentViewType } from '../CustomPropTypes';
