+++ graylog2-web-interface/src/views/logic/fieldtypes/FieldTypeFor.ts
@@ -15,7 +15,7 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 import Series, { isFunction } from 'views/logic/aggregationbuilder/Series';
-import type { FieldTypeMappingsList } from 'views/stores/FieldTypesStore';
+import type { FieldTypeMappingsList } from 'views/logic/fieldtypes/types';
 
 import inferTypeForSeries from './InferTypeForSeries';
 import FieldType from './FieldType';
