+++ graylog2-web-interface/src/views/logic/fieldtypes/types.ts
@@ -1,0 +1,5 @@
+import type * as Immutable from 'immutable';
+
+import type FieldTypeMapping from 'views/logic/fieldtypes/FieldTypeMapping';
+
+export type FieldTypeMappingsList = Immutable.List<FieldTypeMapping>;
